macosx: Use floating window level for effects and other panels
All black-syle panels should get the floating window level.
This is normal behaviour of panels and has the positive side effect
that the panel also stays in front while vout window reconfigurations.

This level is also used for the preferences window, as this is sort
of a modal window which needs to be closed (in order to get settings
saved).

close #12063

(cherry picked from commit 081cba9ccc60b60f5f22a8a5be3895147eb0a709)
Signed-off-by: David Fuhrmann <dfuhrmann@videolan.org>