vdpau/chroma: skip forward to current picture if forced (fixes #11410)
VDPAU introduces a delay of one field (one picture if not deinterlacing).
This is normally not an issue as the PTS is preserved. But it does not
work if there is only one picture not followed by another one.

(cherry picked from commit 6e82ebed93d6f017af4369d8712f7063421ee9ca)

Conflicts:
	modules/hw/vdpau/chroma.c