lua: Do not compile broken metachannels script
The current metachannels script is broken, as the API it uses was
deprecated and it's not trivial to migrate to the new API, which uses
OAuth. This removes it from the makefile so it's not compiled and not
included in VLC, as a broken script is not useful for the user.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit 2bfd9e803684b8d2ff3b1b4b961dd7094f746b35)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>