http: simplify and fix size handling (fixes #17540)
The Content-Range header line indicates the size of the underlying
resource; it has bearings on the size of the response body at the
HTTP protocol level.

Also the start offset in the Content-Range should be ignored, as
overriding it can lead to unexpected situations in VLC. Only a broken
server will not match the requested value anyway.