macosx: Select correct table view item if playlist is in background
If playlist is in background, the data model is not correctly
initialized yet and thus the item is not found. Workaround by
selecting the last index which is always the newly playing item.
This is a hack, but the data model in VLC 2.2.x is broken by design
and is fixed already in VLC 3.0.

Closes #17570