mtp: fix initialization to avoid segfault on close
When an Android device is plugged with MTP disabled, the tracks may not
be retrieved.

In that branch of code, pp_items was never initialized, so the behavior
was undefined in CloseDevice() (on device unplug), often causing a
segfault.

Initialize it to NULL to fix the problem.

Signed-off-by: Romain Vimont <rom@rom1v.com>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit 3ff283bc176b2dd2e83bb7cd7358eb2a0104124c)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>